requirements.txt
alembic==1.13.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi @ file:///private/tmp/certifi-20240307-4874-p3lgkm/certifi-2024.2.2
cffi==1.16.0
charset-normalizer @ file:///private/tmp/python-charset-normalizer-20231123-5225-x00973/charset-normalizer-3.3.2
click==8.1.7
codegen==1.0
cryptography==42.0.6
delocate==0.11.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.2
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna @ file:///private/tmp/python-idna-20231125-5279-xtqzsf/idna-3.6
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
libretranslatepy==2.1.1
lxml==5.2.2
macholib==1.16.3
Mako==1.3.3
MarkupSafe==2.1.5
oauthlib==3.2.2
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pillow==10.3.0
pluggy==1.5.0
polib==1.2.0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
pyjokes==0.6.0
pypdf==4.2.0
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.1.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-translate==1.0.15
PyYAML==6.0.1
requests @ file:///private/tmp/python-requests-20231124-5682-96f89l/requests-2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
tabulate @ file:///private/tmp/python-tabulate-20231211-28538-an4j8r/tabulate-0.9.0
translate==3.6.1
tweepy==4.14.0
typing_extensions==4.11.0
ujson==5.9.0
urllib3 @ file:///private/tmp/python-urllib3-20240218-4962-2z5xut/urllib3-2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.2_1/libexec/wheel-0.42.0-py3-none-any.whl#sha256=5f1ca525bef7466225922019d72cb06a9e91cd3ff807ddd4be1acb84655e5107
